fix: 매핑 충돌 방지를 위해 컨트롤러 파일 변경
