+++ src/main/java/dev/syntax/domain/account/controller/AccountController.java
@@ -8,6 +8,8 @@ import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -17,6 +19,13 @@ import dev.syntax.domain.account.dto.AccountSummaryRes;
 import dev.syntax.domain.account.enums.AccountType;
 import dev.syntax.domain.account.service.AccountHistoryDetailService;
 import dev.syntax.domain.account.service.AccountHistoryService;
+import dev.syntax.domain.card.dto.CardInfoRes;
+import dev.syntax.domain.card.service.CardInquiryService;
+import dev.syntax.domain.transfer.dto.AutoTransferReq;
+import dev.syntax.domain.transfer.dto.AutoTransferRes;
+import dev.syntax.domain.transfer.enums.AutoTransferType;
+import dev.syntax.domain.transfer.service.AutoTransferCreateService;
+import dev.syntax.domain.transfer.service.AutoTransferInquiryService;
 import dev.syntax.domain.account.service.AccountBalanceService;
 import dev.syntax.global.auth.annotation.CurrentUser;
 import dev.syntax.global.auth.dto.UserContext;
@@ -46,6 +55,9 @@ public class AccountController {
         private final AccountHistoryService accountHistoryService;
         private final AccountHistoryDetailService accountHistoryDetailService;
         private final AccountBalanceService accountSummaryService;
+        private final CardInquiryService cardInquiryService;
+        private final AutoTransferInquiryService autoTransferInquiryService;
+        private final AutoTransferCreateService autoTransferCreateService;
 
         /**
          * 📌 [본인] 특정 계좌 잔액 조회
@@ -197,4 +209,69 @@ public class AccountController {
                 return ApiResponseUtil.success(SuccessCode.OK,
                                 accountHistoryDetailService.getDetail(transactionId, user));
         }
+
+            /**
+         * 자녀 본인 카드 조회
+         * GET /account/card
+         */
+        @GetMapping("/card")
+        public ResponseEntity<BaseResponse<?>> getMyCard(@CurrentUser UserContext ctx) {
+                CardInfoRes res = cardInquiryService.getCardInfo(ctx.getId(), ctx);
+                return ApiResponseUtil.success(SuccessCode.OK, res);
+        }
+
+        /**
+         * 부모가 자녀 카드 조회
+         * GET /account/{childId}/card
+         */
+        @GetMapping("/{childId}/card")
+        public ResponseEntity<BaseResponse<?>> getChildCard(
+                @PathVariable Long childId,
+                @CurrentUser UserContext ctx
+        ) {
+                CardInfoRes res = cardInquiryService.getCardInfo(childId, ctx);
+                return ApiResponseUtil.success(SuccessCode.OK, res);
+        }
+
+            /**
+         * 자동이체 설정 조회 API.
+         * <p>
+         * 이 경로는 용돈 자동이체만 접근가능하다는 전제 하에 구현하였습니다.
+         * 자녀의 자동이체 설정 정보를 조회합니다.
+         * </p>
+         *
+         * @param id 자녀 ID (URL 경로 변수)
+         * @param ctx 인증된 사용자 컨텍스트
+         * @return 자동이체 설정 정보 (200 OK)
+         */
+        @GetMapping("/{id}/auto-transfer")
+        public ResponseEntity<BaseResponse<?>> getAutoTransfer(
+                @PathVariable("id") Long id,
+                @CurrentUser UserContext ctx) {
+
+                AutoTransferRes res = autoTransferInquiryService.getAutoTransfer(id, AutoTransferType.ALLOWANCE, ctx);
+                return ApiResponseUtil.success(SuccessCode.OK, res);
+        }
+
+        /**
+         * 자동이체 설정 생성 API.
+         * <p>
+         * 부모가 자녀의 계좌로 자동이체를 설정합니다.
+         * </p>
+         *
+         * @param id 자녀 ID (URL 경로 변수)
+         * @param req 자동이체 설정 요청 정보 (총 금액, 이체일, 투자 비율)
+         * @param ctx 인증된 사용자 컨텍스트 (부모 권한 확인용)
+         * @return 생성 성공 응답 (201 Created)
+         */
+        @PostMapping("/{id}/auto-transfer")
+        public ResponseEntity<BaseResponse<?>> createAutoTransfer(
+                @PathVariable("id") Long id,
+                @RequestBody AutoTransferReq req,
+                @CurrentUser UserContext ctx) {
+
+                autoTransferCreateService.createAutoTransfer(id, req, ctx);
+                return ApiResponseUtil.success(SuccessCode.CREATED);
+        }
+
 }
